+++ src/ArduinoJson/Configuration.hpp
@@ -140,6 +140,11 @@
 #define ARDUINOJSON_ENABLE_NAN 1
 #endif
 
+// Support Infinity in JSON
+#ifndef ARDUINOJSON_ENABLE_INFINITY
+#define ARDUINOJSON_ENABLE_INFINITY 1
+#endif
+
 // Control the exponentiation threshold for big numbers
 // CAUTION: cannot be more that 1e9 !!!!
 #ifndef ARDUINOJSON_POSITIVE_EXPONENTIATION_THRESHOLD